create search input 'to' and set up direction button component
